requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
sentence-transformers